\name{A}
\docType{data}
\alias{A}
\title{Examples of textured images}
\description{A, B, and C are examples of different textured images 
(grayscale). Each matrix object contains an image. Each entry of the 
matrix contains an image intentsity value.}
\usage{data(textures)}
\format{Matricies of varying sizes which each contain a textured image}
\keyword{datasets}
