\name{summary.cddews}
\alias{summary.cddews}
\title{Use summary on a cddews object}
\description{
This function prints out more information about a \code{cddews}  object in a nice readable format.
}
\usage{
\method{summary}{cddews}(object, ...)
}
\arguments{
  \item{object}{An object of class cddews about which you wish to print out more information.}
\item{...}{This argument actually does nothing in this function!}
}
\value{
No value is returned.
}
\references{
Eckley IA, Nason GP, Treloar RL (2010). Locally stationary elds with application to the
modelling and analysis of image texture. Journal of the Royal Statistical Society (Series
C), 59, (to appear).
}
\author{Idris Eckley}
\seealso{\code{\link{cddews}}, \code{\link{print.cddews}}}
\keyword{models}
