% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cddews.R
\docType{methods}
\name{summary.cddews}
\alias{summary.cddews}
\title{Use summary on a cddews object}
\description{
This function prints out more information about a \code{cddews}  object in a nice readable format.
}
\usage{
\method{summary}{cddews}(object, ...)
}
\arguments{
  \item{object}{An object of class cddews about which you wish to print out more information.}
\item{...}{This argument actually does nothing in this function!}
}
\value{
No value is returned.
}
\references{ 
Eckley, I.A., Nason, G.P. and Treloar, R.L. (2010) Locally stationary wavelet fields with application to the modelling and analysis of image texture. Journal of the Royal Statistical Society (Series C), 59, 595 - 616. 

Eckley, I.A. and Nason G.P. (2011). LS2W: Implementing the Locally Stationary 2D Wavelet Process Approach in R, Journal of Statistical Software, 43(3), 1-23.
URL http://www.jstatsoft.org/v43/i03/.
}
\author{Idris Eckley}
\seealso{\code{\link{cddews}}, \code{\link{print.cddews}}}
\keyword{models}
